requirements.txt
azure-functions==1.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
azure-storage-blob>=12.19.0
